Moves request-editor and main-column in their own directories.
