+++ src/app/main-column/main-column.component.css
